+++ io/src/main/java/org/wildfly/extension/io/IOSubsystemParser.java
@@ -22,7 +22,7 @@
  * /
  */
 
-package org.jboss.as.io;
+package org.wildfly.extension.io;
 
 import static org.jboss.as.controller.parsing.ParseUtils.unexpectedElement;
