+++ io/src/main/resources/META-INF/services/org.jboss.as.controller.Extension
@@ -1,2 +1,2 @@
-org.jboss.as.io.IOExtension
+org.wildfly.extension.io.IOExtension
