+++ io/src/test/java/org/wildfly/extension/io/IOSubsystemTestCase.java
@@ -22,7 +22,7 @@
  * /
  */
 
-package org.jboss.as.io;
+package org.wildfly.extension.io;
 
 import java.io.IOException;
